+++ common/src/main/java/org/terraform/cave/NoiseRavine.java
@@ -37,7 +37,7 @@ public class NoiseRavine extends NoiseCaveAbstract{
         float ravine = ravineNoise.GetNoise(3*rawX,y*0.4f,3*rawZ);
         //Multiply by a filter that varies with depth relative to height.
         //At depth 50 blocks, force to 0
-        ravine *= filter*0.5885*Math.log(RAVINE_DEPTH+1-(height-y)); //ASSUMPTION: y <= height.
+        ravine *= (float) (filter*0.5885*Math.log(RAVINE_DEPTH+1-(height-y))); //ASSUMPTION: y <= height.
 //        if(ravine < smallest){
 //            smallest = ravine;
 //            TerraformGeneratorPlugin.logger.info("SMALLEST: " + smallest);
